Add files via upload
Added functionality for ThirtyOne to switch a card in the player's hand from a card in the center. Added new method in Hand to pull the totals if Aces are worth 11 instead of 1. Added new method in Hand to remove a specific card from the hand. Added comments in ThirtyOne to mirror Blackjack's format